refactor: Move Heap & Sentry under frontend/analytics
