+++ src/frontend/analytics/exceptionMonitoring.ts
