+++ src/frontend/analytics/heapAnalytics.ts
